# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a HUB API.",
        "title": "HUB API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/getAccount": {
            "get": {
                "description": "get account information for a specific owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "get account information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account owner address",
                        "name": "owner",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/getBucket": {
            "get": {
                "description": "get bucket information for a specific bucket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bucket"
                ],
                "summary": "get bucket information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "account owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "bucket name",
                        "name": "bucket",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/getNeedle": {
            "get": {
                "description": "get needle information for a specific owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needle"
                ],
                "summary": "get needle information",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "bucket name",
                        "name": "bucket",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "needle name",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/getVolume": {
            "get": {
                "description": "get volume information for a specific owner",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volume"
                ],
                "summary": "get volume information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "file name",
                        "name": "file",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/listAccount": {
            "get": {
                "description": "get a list of accounts with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "list accounts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "get a list of accounts with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "list accounts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/listBucket": {
            "get": {
                "description": "get a list of buckets with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bucket"
                ],
                "summary": "list buckets",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "get a list of buckets with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bucket"
                ],
                "summary": "list buckets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address",
                        "name": "owner",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/listNeedle": {
            "get": {
                "description": "get a list of needles for a specific owner with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needle"
                ],
                "summary": "list needles",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "bucket name",
                        "name": "bucket",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "get a list of needles for a specific owner with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "needle"
                ],
                "summary": "list needles",
                "parameters": [
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "owner address",
                        "name": "owner",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "\"\"",
                        "description": "bucket name",
                        "name": "bucket",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
        "/api/listVolume": {
            "get": {
                "description": "get a list of volumes for a specific owner with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volume"
                ],
                "summary": "list volumes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address",
                        "name": "owner",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "get a list of volumes for a specific owner with pagination support",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "volume"
                ],
                "summary": "list volumes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "owner address",
                        "name": "owner",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "pagination offset",
                        "name": "offset",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "default": 32,
                        "description": "number of items per page",
                        "name": "length",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "599": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/error.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "error.APIError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    }
}